Fixes. Fix databases updating issued timestamps counter logic. A time stamp request body should be also be encoded in Base64. And a few little fixes.

+++ picky-server/src/db/mongodb/model.rs
@@ -154,7 +154,7 @@ impl mongodm::CollectionConfig for TimestampCollConf {
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct IssuedTimestampsCounter {
-    pub counter: u32,
+    pub counter: i64,
 }
 
 impl mongodm::Model for IssuedTimestampsCounter {
